requirements.txt
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
bitarray==3.1.1
black==25.1.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
ckzg==2.0.1
click==8.1.8
cytoolz==1.0.1
dataclasses-json==0.6.7
distro==1.9.0
eth-account==0.13.5
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.6.1
eth-rlp==2.2.0
eth-typing==5.2.0
eth-utils==5.2.0
eth_abi==5.2.0
frozenlist==1.5.0
google-auth==2.38.0
google-auth-oauthlib==1.2.1
gspread==6.2.0
gspread-dataframe==4.0.0
h11==0.14.0
hexbytes==1.3.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.20
langchain-community==0.3.19
langchain-core==0.3.41
langchain-text-splitters==0.3.6
langsmith==0.3.12
marshmallow==3.26.1
multidict==6.1.0
mypy-extensions==1.0.0
numpy==2.2.3
oauth2client==4.1.3
oauthlib==3.2.2
openai==1.65.4
orjson==3.10.15
packaging==24.2
pandas==2.2.3
parsimonious==0.10.0
pathspec==0.12.1
platformdirs==4.3.6
propcache==0.3.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycryptodome==3.21.0
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
pyunormalize==16.0.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rlp==4.1.0
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
tenacity==9.0.0
toolz==1.0.0
tqdm==4.67.1
types-requests==2.32.0.20250306
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
watchdog==6.0.0
web3==7.8.0
websockets==13.1
yarl==1.18.3
zstandard==0.23.0
